models/svc/vevo2/requirements.txt
praat-parselmouth
torchcrepe

openai-whisper
